Rename internal/pki.go to internal/pki/policy_identifier.go
